+++ packages/admin/cms-admin/src/blocks/types.tsx
@@ -79,6 +79,7 @@ export interface BlockMethods<
     dependencies?: (state: State) => BlockDependency[];
     replaceDependenciesInOutput: (output: OutputApi, replacements: ReplaceDependencyObject[]) => OutputApi;
     resolveDependencyPath: (state: State, jsonPath: string) => string;
+    extractTextContents?: (state: State, options: { includeInvisibleContent: boolean }) => string[];
 }
 
 export interface AnonymousBlockInterface<
@@ -93,6 +94,7 @@ export interface AnonymousBlockInterface<
     AdminComponent: BlockAdminComponent<State>;
     definesOwnPadding?: boolean;
     definesOwnTitle?: boolean;
+    extractTextContents?: BlockMethods<InputApi, State, OutputApi, PreviewState>["extractTextContents"];
 }
 
 export interface BlockInterface<