+++ packages/admin/admin-date-time/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @comet/admin-date-time
 
+## 7.14.0
+
+### Patch Changes
+
+- Updated dependencies [6b75f20e4]
+    - @comet/admin@7.14.0
+    - @comet/admin-icons@7.14.0
+
 ## 7.13.0
 
 ### Patch Changes